Added security calls to account routes [ST-44]

+++ server/src/main/kotlin/com/studystream/app/server/feature/account/routes/Delete.kt
@@ -1,9 +1,13 @@
 package com.studystream.app.server.feature.account.routes
 
+import com.studystream.app.domain.model.Account
+import com.studystream.app.domain.security.Permission
 import com.studystream.app.domain.service.AccountService
 import com.studystream.app.server.feature.account.Accounts
+import com.studystream.app.server.security.requireAccount
+import com.studystream.app.server.security.requirePermission
 import com.studystream.app.server.utils.endpoint
-import com.studystream.app.server.utils.typeSafeGet
+import com.studystream.app.server.utils.typeSafeDelete
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
@@ -14,9 +18,10 @@ import org.koin.ktor.ext.get
 
 internal fun Routing.installDeleteAccountsRoute() {
     authenticate {
-        typeSafeGet<Accounts.AccountId> { route ->
+        typeSafeDelete<Accounts.AccountId> { route ->
             deleteAccount(
                 route = route,
+                account = call.requireAccount(),
                 accountService = call.get(),
             )
 
@@ -27,9 +32,11 @@ internal fun Routing.installDeleteAccountsRoute() {
 
 suspend fun deleteAccount(
     route: Accounts.AccountId,
+    account: Account,
     accountService: AccountService,
 ) = endpoint {
-    // TODO: add admin check
+    account.requirePermission(Permission.DELETE_ACCOUNTS)
+
     accountService
         .deleteAccount(route.id)
         .getOrThrow()
